requirements.txt
pillow==10.1.0
tk==0.1.0
moviepy==1.0.3
pygame==2.5.2
